+++ frontend/components/YotpoMainHtmlReferenceContainer/components/YotpoMainWidget/index.jsx
@@ -4,9 +4,9 @@ import styles from './style';
 import getYotpo from '../../../../helpers/getYotpo';
 
 /**
- * Mounts yotpo script.
+ * Yotpo main widget
  */
-class YotpoScript extends Component {
+class YotpoMainWidget extends Component {
   static propTypes = {
     config: PropTypes.shape().isRequired,
     productId: PropTypes.string.isRequired,
@@ -19,7 +19,7 @@ class YotpoScript extends Component {
 
   /**
    * Constructs
-   * @param {Objects} props props
+   * @param {Object} props props
    */
   constructor(props) {
     super(props);
@@ -34,8 +34,8 @@ class YotpoScript extends Component {
       getYotpo(this.props.config.yotpoAppKey);
       return;
     }
-    window.yotpo.initialized = false;
     window.yotpo.clean();
+    window.yotpo.refreshWidgets();
     this.yotpoMountTimeout = setTimeout(() => {
       window.yotpo.initWidgets();
       this.yotpoMountTimeout = undefined;
@@ -46,6 +46,8 @@ class YotpoScript extends Component {
    * Cleanup timeout request
    */
   componentWillUnmount() {
+    // Set initialized to false on unmounting
+    window.yotpo.initialized = false;
     clearTimeout(this.yotpoMountTimeout);
   }
 
@@ -58,28 +60,16 @@ class YotpoScript extends Component {
       return null;
     }
     return (
-      <div>
-        <div
-          className={`yotpo bottomLine ${styles.padding}`}
-          data-appkey={`${this.props.config.yotpoAppKey}`}
-          data-domain=""
-          data-product-id={`${this.props.productId}`}
-          data-name={`${this.props.product.name}`}
-          data-url={`${this.props.product.productUrl}`}
-          data-image-url={`${this.props.product.featuredImageUrl}`}
-          data-description=""
-        />
-        <div
-          className={`yotpo yotpo-main-widget ${styles.container} ${styles.padding}`}
-          data-product-id={`${this.props.productId}`}
-          data-name={`${this.props.product.name}`}
-          data-url={`${this.props.product.productUrl}`}
-          data-image-url={`${this.props.product.featuredImageUrl}`}
-          data-description=""
-        />
-      </div>
+      <div
+        className={`yotpo yotpo-main-widget ${styles.container} ${styles.padding}`}
+        data-product-id={`${this.props.productId}`}
+        data-name={`${this.props.product.name}`}
+        data-url={`${this.props.product.productUrl}`}
+        data-image-url={`${this.props.product.featuredImageUrl}`}
+        data-description=""
+      />
     );
   }
 }
 
-export default YotpoScript;
+export default YotpoMainWidget;