+++ CHANGELOG.md
@@ -2,6 +2,11 @@
  All notable changes to this project will be documented in this file.
  The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.0] - 2019-04-30
+- Added style rule to hide yotpo product slider
+- Added aggregated star review in product header with scroll to functionality
+- Updated potential bug to refresh widgets on unmounting.
+
 ## [1.0.3] - 2019-03-21
 Fixed a typo.
 
@@ -18,6 +23,7 @@ First version of the extension.
 - You can configure a Yotpo app key for specific merchant.
 - Connect to write a Review is not included in this integration.
 
+[1.1.0]: https://github.com/shopgate-professional-services/ext-yotpo-reviews/compare/v1.0.3...v1.1.0
 [1.0.3]: https://github.com/shopgate-professional-services/ext-yotpo-reviews/compare/v1.0.2...v1.0.3
 [1.0.2]: https://github.com/shopgate-professional-services/ext-yotpo-reviews/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/shopgate-professional-services/ext-yotpo-reviews/compare/v1.0.0...v1.0.1